Refactored Properties.  Changed AddStatChanges.
Seeing as how I refactored PropertiesBase to
Properties, to prevent conflict I refactored the
variable Properties in this to properties.

I Changed AddStatChanges so that it would return a
UnitStats and the UnitStats passed in is not passed
in by Ref.
This is mainly to conform with how, in previous
commits, I changed how we go about adding UnitStats
Together.
Instead of using the += operator, now we return the
result of UnitStats.Add.