Made Derive from UnitProperties
I made this derive from UnitProperties for reasons I
litsed on the commit for UnitProperties.
Because of this, I removed Stats, as it is not needed
.

Also, I changed passive into Effects, which is an
array.  This is to go along with the decision that
passives should be called Effects, and that there is
no reason why there might not be multiple effects for
one champion.