Changed a lot.
First thing, I added a layout for
the client telling the server if
and which ability they want to use.
Currently does nothing.

Second, I made ClientAxis check if
the client is dead before moving.
This is because dead things don't
move. (:P)

Third, I made a changes so that the
Move Speed and Rotate Speed of the
player is how much they will move/
rotate in a second.  This is to make
 that fps independent.

Fourth, I removed the RPC functions
Freeze, and Unfreeze because they are
 sloppy and unused.

Fifth, I changed the Spectate
functions to fix some bugs.